update onboarding screen structure
