--- modules/react-native-onboarding-v2/index.js
@@ -1,13 +1,0 @@
-import * as React from 'react';
-import Slider from './components/Slider/index'
-import options from './options'
-
-const Onboarding = ()=> {
-  return (
-    <Slider data={options.sliderData} onFinish={() => {}}/>
-  );
-}
-export default {
-  title: "Onboarding",
-  navigator: Onboarding
-}
