--- modules/react-native-onboarding-v2/options.js
@@ -1,20 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const sliderData = [{
-  heading: 'Lorem ipsum',
-  description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.tortor lectus tempus lacus.'
-}, {
-  heading: 'Lorem ipsum 1',
-  description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.tortor lectus tempus lacus 1.'
-},{
-  heading: 'Lorem ipsum 2',
-  description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.tortor lectus tempus lacus 2.'
-}]
-const styles = StyleSheet.create({
-  
-});
-
-export default {
-  styles: styles,
-  sliderData: sliderData
-}
